fix: move bulk action components to file-row to fix max-lines lint
