+++ apps/web/components/task/changes-panel-file-row.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useMemo } from "react";
 import {
   IconArrowBackUp,
   IconPlus,
@@ -9,6 +10,7 @@ import {
   IconPencil,
 } from "@tabler/icons-react";
 
+import { Button } from "@kandev/ui/button";
 import { cn } from "@/lib/utils";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@kandev/ui/tooltip";
 import { LineStat } from "@/components/diff-stat";
@@ -196,6 +198,119 @@ function FileRowActions({
         </TooltipTrigger>
         <TooltipContent>Edit</TooltipContent>
       </Tooltip>
+    </div>
+  );
+}
+
+// --- Bulk action components (used by FileListSection) ---
+
+export function DefaultActionButtons({
+  actionLabel,
+  isActionLoading,
+  onAction,
+  secondaryActionLabel,
+  isSecondaryActionLoading,
+  onSecondaryAction,
+}: {
+  actionLabel: string;
+  isActionLoading?: boolean;
+  onAction: () => void;
+  secondaryActionLabel?: string;
+  isSecondaryActionLoading?: boolean;
+  onSecondaryAction?: () => void;
+}) {
+  return (
+    <>
+      <Button
+        size="sm"
+        variant="outline"
+        className="h-6 text-[11px] px-2.5 gap-1 cursor-pointer"
+        onClick={onAction}
+        disabled={isActionLoading}
+      >
+        {isActionLoading && <IconLoader2 className="h-3 w-3 animate-spin" />}
+        {actionLabel}
+      </Button>
+      {onSecondaryAction && secondaryActionLabel && (
+        <Button
+          size="sm"
+          variant="outline"
+          className="h-6 text-[11px] px-2.5 gap-1 cursor-pointer"
+          onClick={onSecondaryAction}
+          disabled={isSecondaryActionLoading}
+        >
+          {isSecondaryActionLoading && <IconLoader2 className="h-3 w-3 animate-spin" />}
+          {secondaryActionLabel}
+        </Button>
+      )}
+    </>
+  );
+}
+
+export function BulkActionBar({
+  variant,
+  selectionCount,
+  selectedPaths,
+  onBulkStage,
+  onBulkUnstage,
+  onBulkDiscard,
+  onClearSelection,
+}: {
+  variant: "unstaged" | "staged";
+  selectionCount: number;
+  selectedPaths: Set<string>;
+  onBulkStage?: (paths: string[]) => void;
+  onBulkUnstage?: (paths: string[]) => void;
+  onBulkDiscard?: (paths: string[]) => void;
+  onClearSelection: () => void;
+}) {
+  const paths = useMemo(() => [...selectedPaths], [selectedPaths]);
+
+  return (
+    <div data-testid={`bulk-actions-${variant}`} className="flex items-center gap-1.5">
+      <span className="text-[11px] text-muted-foreground">{selectionCount} selected</span>
+      {variant === "unstaged" && onBulkStage && (
+        <Button
+          data-testid="bulk-stage"
+          size="sm"
+          variant="outline"
+          className="h-6 text-[11px] px-2.5 gap-1 cursor-pointer"
+          onClick={() => {
+            onBulkStage(paths);
+            onClearSelection();
+          }}
+        >
+          Stage {selectionCount}
+        </Button>
+      )}
+      {variant === "staged" && onBulkUnstage && (
+        <Button
+          data-testid="bulk-unstage"
+          size="sm"
+          variant="outline"
+          className="h-6 text-[11px] px-2.5 gap-1 cursor-pointer"
+          onClick={() => {
+            onBulkUnstage(paths);
+            onClearSelection();
+          }}
+        >
+          Unstage {selectionCount}
+        </Button>
+      )}
+      {onBulkDiscard && (
+        <Button
+          data-testid="bulk-discard"
+          size="sm"
+          variant="outline"
+          className="h-6 text-[11px] px-2.5 gap-1 cursor-pointer text-destructive hover:text-destructive"
+          onClick={() => {
+            onBulkDiscard(paths);
+            onClearSelection();
+          }}
+        >
+          Discard {selectionCount}
+        </Button>
+      )}
     </div>
   );
 }
